requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
httpx==0.28.1
pydantic==2.11.3
aiosqlite==0.20.0
python-dotenv==1.0.1
streamlit==1.44.1
crewai
crewai_tools
